Trait nutsfinance_stable_asset::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type AssetId: Parameter + Ord + Copy;
    type Balance: Parameter + Codec + Copy + Ord + From<Self::AtLeast64BitUnsigned> + Zero;
    type Assets: Inspect<Self::AccountId, AssetId = Self::AssetId, Balance = Self::Balance> + Mutate<Self::AccountId, AssetId = Self::AssetId, Balance = Self::Balance> + Transfer<Self::AccountId, AssetId = Self::AssetId, Balance = Self::Balance>;
    type AtLeast64BitUnsigned: Parameter + CheckedAdd + CheckedSub + CheckedMul + CheckedDiv + Copy + Eq + Ord + From<Self::Balance> + From<u8> + From<u128> + From<Self::BlockNumber> + TryFrom<usize> + Zero + One + FixedPointOperand;
    type PalletId: Get<PalletId>;
    type FeePrecision: Get<Self::AtLeast64BitUnsigned>;
    type SwapExactOverAmount: Get<Self::AtLeast64BitUnsigned>;
    type APrecision: Get<Self::AtLeast64BitUnsigned>;
    type PoolAssetLimit: Get<u32>;
    type WeightInfo: WeightInfo;
    type EnsurePoolAssetId: ValidateAssetId<Self::AssetId>;
    type ListingOrigin: EnsureOrigin<Self::Origin>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type AssetId: Parameter + Ord + Copy

type Balance: Parameter + Codec + Copy + Ord + From<Self::AtLeast64BitUnsigned> + Zero

type Assets: Inspect<Self::AccountId, AssetId = Self::AssetId, Balance = Self::Balance> + Mutate<Self::AccountId, AssetId = Self::AssetId, Balance = Self::Balance> + Transfer<Self::AccountId, AssetId = Self::AssetId, Balance = Self::Balance>

type AtLeast64BitUnsigned: Parameter + CheckedAdd + CheckedSub + CheckedMul + CheckedDiv + Copy + Eq + Ord + From<Self::Balance> + From<u8> + From<u128> + From<Self::BlockNumber> + TryFrom<usize> + Zero + One + FixedPointOperand

type PalletId: Get<PalletId>

type FeePrecision: Get<Self::AtLeast64BitUnsigned>

type SwapExactOverAmount: Get<Self::AtLeast64BitUnsigned>

type APrecision: Get<Self::AtLeast64BitUnsigned>

type PoolAssetLimit: Get<u32>

type WeightInfo: WeightInfo

type EnsurePoolAssetId: ValidateAssetId<Self::AssetId>

type ListingOrigin: EnsureOrigin<Self::Origin>

The origin which may create pool or modify pool.